+++ Common/ProjectileCommon/Abstract/ChargeBowProjectile.cs
@@ -4,9 +4,9 @@ using System.IO;
 using SpiritReforged.Common.PlayerCommon;
 using SpiritReforged.Common.Easing;
 
-namespace SpiritReforged.Common.ProjectileCommon;
+namespace SpiritReforged.Common.ProjectileCommon.Abstract;
 
-public abstract class BaseChargeBow(float maxChargePower = 2f, float perfectShotPower = 1.5f, int perfectShotTime = 30) : ModProjectile
+public abstract class ChargeBowProjectile(float maxChargePower = 2f, float perfectShotPower = 1.5f, int perfectShotTime = 30) : ModProjectile
 {
 	private const int STRING_BOUNCE_TIME = 30; //Could be adjusted to be dynamic if really needed
 
@@ -71,7 +71,6 @@ public abstract class BaseChargeBow(float maxChargePower = 2f, float perfectShot
 		player.SetCompositeArmBack(true, Player.CompositeArmStretchAmount.Full, player.itemRotation);
 
 		if (Main.myPlayer == Projectile.owner)
-		{
 			if (!_fired)
 			{
 				Projectile.timeLeft = STRING_BOUNCE_TIME;
@@ -91,7 +90,6 @@ public abstract class BaseChargeBow(float maxChargePower = 2f, float perfectShot
 					Projectile.netUpdate = true;
 				}
 			}
-		}
 
 		if (_fired)
 			AfterShoot();
