+++ Common/WorldGeneration/Microbiomes/Biomes/Ziggurat/ZigguratMicrobiome.Infestation.cs
@@ -7,7 +7,7 @@ using SpiritReforged.Content.Ziggurat.Walls;
 
 namespace SpiritReforged.Common.WorldGeneration.Microbiomes.Biomes.Ziggurat;
 
-public partial class ZigguratMicrobiome : Microbiome
+public partial class ZigguratMicrobiome : MicrobiomeSystem.Microbiome
 {
 	private static void Infest(int count, Rectangle bound)
 	{
