+++ Common/WorldGeneration/Microbiomes/Microbiome.cs
@@ -5,74 +5,71 @@ using Terraria.ModLoader.IO;
 
 namespace SpiritReforged.Common.WorldGeneration.Microbiomes;
 
-/// <summary> Can be used to contain data for individually-generated biome instances. </summary>
-public abstract class Microbiome : ILoadable
-{
-	public virtual string Name => GetType().Name;
-	public Point16 Position { get; private set; }
-
-	/// <summary> Creates a microbiome of the provided type and places it according to <paramref name="point"/> and <paramref name="place"/>. </summary>
-	/// <param name="point"> The origin position of the microbiome. </param>
-	/// <param name="place"> Whether this instance should be generated at <paramref name="point"/>. </param>
-	/// <returns> The newly created instance. </returns>
-	public static T Create<T>(Point16 point, bool place = true) where T : Microbiome
-	{
-		var inst = MicrobiomeSystem.GetInstance<T>();
-		inst.Position = point;
-
-		if (place)
-			inst.OnPlace(point);
-
-		MicrobiomeSystem.Microbiomes.Add(inst);
-		return inst;
-	}
-
-	/// <inheritdoc cref="Create{T}(Point16, bool)"/>
-	public static T Create<T>(Point point, bool place = true) where T : Microbiome => Create<T>(new Point16(point.X, point.Y), place);
-
-	protected abstract void OnPlace(Point16 point);
-
-	/// <summary> Can be used to save custom data related to this microbiome <b>instance</b>. <para/>
-	/// <see cref="Position"/> is saved by default. If that's not necessary, override this method. </summary>
-	public virtual void WorldSave(TagCompound tag)
-	{
-		if (Position != Point16.Zero)
-			tag[nameof(Position)] = Position; //Don't write the zeroed value because that's a waste
-	}
-
-	public virtual void WorldLoad(TagCompound tag) => Position = tag.Get<Point16>(nameof(Position));
-
-	public virtual void NetSend(BinaryWriter writer) => writer.WritePoint16(Position);
-	public virtual void NetReceive(BinaryReader reader) => Position = reader.ReadPoint16();
-
-	public void Load(Mod mod)
-	{
-		MicrobiomeSystem.AddDefinition(this);
-		Load();
-	}
-
-	/// <summary> Called once per type when <see cref="MicrobiomeSystem.BiomeByName"/> is populated. </summary>
-	public virtual void Load() { }
-	public virtual void Unload() { }
-
-	/// <summary> Simply casts the result of <see cref="object.MemberwiseClone"/>. </summary>
-	public Microbiome Clone() => (Microbiome)MemberwiseClone();
-}
-
 public sealed class MicrobiomeSystem : ModSystem
 {
+	/// <summary> Data structure for individually-generated biome instances that are automatically loaded and synced in multiplayer. <para/>
+	/// Contains <see cref="Position"/> by default and can be placed using method <code>Create(Point16 | Point)</code> </summary>
+	public abstract class Microbiome : ILoadable
+	{
+		public virtual string Name => GetType().Name;
+		public Point16 Position { get; private set; }
+
+		/// <summary> Creates a microbiome of the provided type at <paramref name="point"/>. </summary>
+		/// <param name="point"> The origin position of the microbiome. </param>
+		/// <returns> The newly created microbiome instance. </returns>
+		public static T Create<T>(Point16 point) where T : Microbiome
+		{
+			var instance = (T)_biomeByName[typeof(T).Name].Clone();
+			instance.Position = point;
+
+			Microbiomes.Add(instance);
+			return instance;
+		}
+
+		/// <inheritdoc cref="Create{T}(Point16)"/>
+		public static T Create<T>(Point point) where T : Microbiome => Create<T>(new Point16(point.X, point.Y));
+
+		/// <summary> Can be used to save custom data related to this microbiome <b>instance</b>. <para/>
+		/// <see cref="Position"/> is saved by default. If that's not necessary, override this method. </summary>
+		public virtual void WorldSave(TagCompound tag)
+		{
+			if (Position != Point16.Zero)
+				tag[nameof(Position)] = Position; //Don't write the zeroed value because that's a waste
+		}
+
+		public virtual void WorldLoad(TagCompound tag) => Position = tag.Get<Point16>(nameof(Position));
+
+		public virtual void NetSend(BinaryWriter writer) => writer.WritePoint16(Position);
+		public virtual void NetReceive(BinaryReader reader) => Position = reader.ReadPoint16();
+
+		public void Load(Mod mod)
+		{
+			AddDefinition(this);
+			Load();
+		}
+
+		/// <summary> Called once per type when <see cref="_biomeByName"/> is populated. </summary>
+		public virtual void Load() { }
+		public virtual void Unload() { }
+
+		/// <summary> Simply casts the result of <see cref="object.MemberwiseClone"/>. </summary>
+		public Microbiome Clone() => (Microbiome)MemberwiseClone();
+	}
+
 	/// <summary> Invoked directly after <see cref="Microbiomes"/> is populated, which can happen during world load or after microbiomes are synced in multiplayer. </summary>
 	public static event Action PopulateMicrobiomes;
 
 	/// <summary> Default microbiome definitions by name, added during load. Instances should not be used directly but instead cloned using <see cref="Microbiome.Clone"/>. </summary>
-	private static readonly Dictionary<string, Microbiome> BiomeByName = [];
+	private static readonly Dictionary<string, Microbiome> _biomeByName = [];
 	/// <summary> All Microbiome instances that currently exist in the world. </summary>
-	internal static readonly List<Microbiome> Microbiomes = [];
+	public static readonly List<Microbiome> Microbiomes = [];
 
-	public static void AddDefinition(Microbiome biome) => BiomeByName.Add(biome.Name, biome);
-	/// <summary> Gets a cloned instance from <see cref="BiomeByName"/>. <para/>
+	/// <summary> Registers this biome instance as a template. </summary>
+	public static void AddDefinition(Microbiome biome) => _biomeByName.Add(biome.Name, biome);
+
+	/*/// <summary> Gets a cloned instance from <see cref="BiomeByName"/>. <para/>
 	/// Prefer <see cref="Microbiome.Create"/> as it automatically registers an instance to <see cref="Microbiomes"/>. </summary>
-	public static T GetInstance<T>() where T : Microbiome => (T)BiomeByName[typeof(T).Name].Clone();
+	public static T GetInstance<T>() where T : Microbiome => (T)BiomeByName[typeof(T).Name].Clone();*/
 
 	public override void ClearWorld() => Microbiomes.Clear();
 
@@ -96,7 +93,7 @@ public sealed class MicrobiomeSystem : ModSystem
 		{
 			string name = reader.ReadString();
 
-			var biome = BiomeByName[name].Clone();
+			Microbiome biome = _biomeByName[name].Clone();
 			biome.NetReceive(reader);
 
 			Microbiomes.Add(biome); //Repopulate Microbiomes based on data provided by the server
@@ -139,9 +136,9 @@ public sealed class MicrobiomeSystem : ModSystem
 			string name = item.GetString("name");
 			TagCompound data = item.GetCompound("data");
 
-			if (BiomeByName.TryGetValue(name, out Microbiome b))
+			if (_biomeByName.TryGetValue(name, out Microbiome b))
 			{
-				var inst = b.Clone();
+				Microbiome inst = b.Clone();
 				inst.WorldLoad(data);
 				Microbiomes.Add(inst);
 			}
